[raft/scd] Extract update and create constraint
